+++ R/template.R
@@ -7,10 +7,6 @@ template_tabularray <- function(theme = "default") {
     out <- readLines(
       system.file("templates/tabularray_default.tex", package = "tinytable")
     )
-  } else if (theme == "bootstrap") {
-    out <- readLines(
-      system.file("templates/tabularray_bootstrap.tex", package = "tinytable")
-    )
   } else if (theme == "striped") {
     out <- readLines(
       system.file("templates/tabularray_default.tex", package = "tinytable")